requirements.txt
joblib~=1.4.0
matplotlib~=3.8.4
numpy~=1.26.4
optuna~=3.6.1
plotly~=5.21.0
scikit-learn~=1.4.2
Pillow~=10.3.0
setuptools